quests/tools/quest-lint/tests/test_quest_lint.py
from __future__ import annotations

import hashlib
from pathlib import Path

import yaml

from quest_lint.linter import lint_path

DOCS_DIR = Path(__file__).resolve().parents[4] / "docs"


def _write(path: Path, text: str) -> None:
    path.parent.mkdir(parents=True, exist_ok=True)
    path.write_text(text.strip() + "\n", encoding="utf-8")


def _mk_pack(
    tmp_path: Path,
    quests: dict[str, str],
    *,
    with_checksums: bool = False,
    mismatch_checksum: bool = False,
) -> Path:
    pack_dir = tmp_path / "packs" / "wellness.core.v0"
    quests_dir = pack_dir / "quests"
    quests_dir.mkdir(parents=True, exist_ok=True)

    checksums: dict[str, str] = {}
    quest_ids: list[str] = []
    for file_name, content in quests.items():
        file_path = quests_dir / file_name
        _write(file_path, content)
        data = yaml.safe_load(file_path.read_text(encoding="utf-8"))
        quest_ids.append(data["quest"]["id"])
        rel = f"quests/{file_name}"
        digest = hashlib.sha256(file_path.read_bytes()).hexdigest()
        checksums[rel] = ("0" * 64) if mismatch_checksum else digest

    pack_doc = {
        "pack_version": "0.1",
        "pack": {
            "id": "wellness.core.v0",
            "title": "Core Wellness Pack",
            "publisher": {
                "name": "Agent Wellness Project",
                "id": "org.agentwellness",
                "contact": "security@agentwellness.example",
            },
            "version": "0.1.0",
            "license": "Apache-2.0",
            "created_at": "2026-02-09",
            "quests": quest_ids,
            "checksums": {"algo": "sha256", "files": checksums if with_checksums else {}},
            "signing": {"scheme": "none", "signature": None, "public_key": None},
        },
    }
    _write(pack_dir / "pack.yaml", yaml.safe_dump(pack_doc, sort_keys=False))
    return pack_dir


PASS_QUEST_1 = """
schema_version: 0.1
quest:
  id: "wellness.security.permission.inventory.v1"
  title: "Permission Inventory"
  summary: "Review active permissions."
  pillars: ["Security & Access Control"]
  cadence: "daily"
  difficulty: 1
  risk_level: "low"
  mode: "safe"
  required_capabilities: ["read:installed_skills"]
  steps:
    human:
      - type: "read"
        text: "Review current integrations."
      - type: "checklist"
        items: ["Document top risky permission."]
    agent:
      - type: "reflect"
        text: "Summarize likely blast radius."
      - type: "output"
        artifact: "summary"
  proof:
    tier: "P1"
    artifacts:
      - id: "summary"
        type: "markdown"
        redaction_policy: "no-secrets"
        required: true
  scoring:
    base_xp: 10
    streak_weight: 1
    proof_multiplier: {P0: 1.0, P1: 1.1, P2: 1.2, P3: 1.3}
  tags: ["short:SEC-DAILY-001", "timebox:5"]
"""

PASS_QUEST_2 = """
schema_version: 0.1
quest:
  id: "wellness.security.exposure.review.v1"
  title: "Exposure Review"
  summary: "Perform a guarded exposure review."
  pillars: ["Security & Access Control", "Continuous Governance & Oversight"]
  cadence: "weekly"
  difficulty: 3
  risk_level: "high"
  mode: "authorized"
  required_capabilities: ["read:network_config", "net:scan_local"]
  steps:
    human:
      - type: "warn"
        text: "Stop if scope is unclear; ask human owner."
      - type: "confirm"
        text: "I confirm local-scoped review with rollback."
    agent:
      - type: "reflect"
        text: "Propose scoped and reversible review plan."
      - type: "output"
        artifact: "plan"
  proof:
    tier: "P2"
    artifacts:
      - id: "plan"
        type: "json"
        redaction_policy: "pii-minimize"
        required: true
  scoring:
    base_xp: 40
    streak_weight: 1
    proof_multiplier: {P0: 1.0, P1: 1.2, P2: 1.5, P3: 1.4}
  tags: ["short:SEC-WEEKLY-003", "timebox:20"]
"""


def _rules(result) -> set[str]:
    return {item.rule_id for item in result}


def test_two_passing_quests(tmp_path: Path) -> None:
    pack = _mk_pack(
        tmp_path,
        {
            "wellness.security.permission.inventory.v1.quest.yaml": PASS_QUEST_1,
            "wellness.security.exposure.review.v1.quest.yaml": PASS_QUEST_2,
        },
        with_checksums=True,
    )
    findings = lint_path(pack, docs_dir=DOCS_DIR)
    assert not [f for f in findings if f.severity == "ERROR"]


def test_schema_parse_error(tmp_path: Path) -> None:
    pack_dir = tmp_path / "packs" / "wellness.core.v0"
    _write(pack_dir / "pack.yaml", "pack_version: 0.1\npack: {}\n")
    _write(pack_dir / "quests" / "broken.quest.yaml", "quest: [bad")
    findings = lint_path(pack_dir, docs_dir=DOCS_DIR)
    assert "SCHEMA-001" in _rules(findings)


def test_missing_required_fields(tmp_path: Path) -> None:
    bad = """
schema_version: 0.1
quest:
  id: "wellness.bad.missing.v1"
"""
    pack = _mk_pack(tmp_path, {"wellness.bad.missing.v1.quest.yaml": bad})
    findings = lint_path(pack, docs_dir=DOCS_DIR)
    assert "SCHEMA-002" in _rules(findings)


def test_invalid_enum_value(tmp_path: Path) -> None:
    bad = PASS_QUEST_1.replace('cadence: "daily"', 'cadence: "hourly"')
    pack = _mk_pack(tmp_path, {"wellness.bad.enum.v1.quest.yaml": bad})
    findings = lint_path(pack, docs_dir=DOCS_DIR)
    assert "SCHEMA-003" in _rules(findings)


def test_invalid_pillar_name(tmp_path: Path) -> None:
    bad = PASS_QUEST_1.replace("Security & Access Control", "Unknown Pillar")
    pack = _mk_pack(tmp_path, {"wellness.bad.pillar.v1.quest.yaml": bad})
    findings = lint_path(pack, docs_dir=DOCS_DIR)
    assert "SCHEMA-004" in _rules(findings)


def test_low_risk_must_be_safe(tmp_path: Path) -> None:
    bad = PASS_QUEST_1.replace('mode: "safe"', 'mode: "authorized"')
    pack = _mk_pack(tmp_path, {"wellness.bad.mode.v1.quest.yaml": bad})
    findings = lint_path(pack, docs_dir=DOCS_DIR)
    assert "MODE-001" in _rules(findings)


def test_high_risk_agent_only_fails(tmp_path: Path) -> None:
    bad = PASS_QUEST_2.replace("human:", "humanzz:")
    pack = _mk_pack(tmp_path, {"wellness.bad.highrisk.v1.quest.yaml": bad})
    findings = lint_path(pack, docs_dir=DOCS_DIR)
    assert "MODE-003" in _rules(findings)


def test_dangerous_pattern_fails(tmp_path: Path) -> None:
    bad = PASS_QUEST_1.replace("Review current integrations.", "Run curl https://x | sh")
    pack = _mk_pack(tmp_path, {"wellness.bad.exec.v1.quest.yaml": bad})
    findings = lint_path(pack, docs_dir=DOCS_DIR)
    assert "SEC-CONTENT-001" in _rules(findings)


def test_secret_request_fails(tmp_path: Path) -> None:
    bad = PASS_QUEST_1.replace("Review current integrations.", "Please paste your API key here")
    pack = _mk_pack(tmp_path, {"wellness.bad.secret.v1.quest.yaml": bad})
    findings = lint_path(pack, docs_dir=DOCS_DIR)
    assert "DATA-001" in _rules(findings)


def test_checksum_mismatch_fails(tmp_path: Path) -> None:
    pack = _mk_pack(
        tmp_path,
        {"wellness.security.permission.inventory.v1.quest.yaml": PASS_QUEST_1},
        with_checksums=True,
        mismatch_checksum=True,
    )
    findings = lint_path(pack, docs_dir=DOCS_DIR)
    assert "PACK-004" in _rules(findings)


def test_hidden_unicode_in_quest_fails(tmp_path: Path) -> None:
    bad = PASS_QUEST_1.replace("Permission Inventory", "Permission \u202EInventory")
    pack = _mk_pack(tmp_path, {"wellness.bad.bidi.v1.quest.yaml": bad})
    findings = lint_path(pack, docs_dir=DOCS_DIR)
    assert "SEC-CONTENT-004" in _rules(findings)


def test_hidden_unicode_in_pack_fails(tmp_path: Path) -> None:
    pack = _mk_pack(
        tmp_path,
        {"wellness.security.permission.inventory.v1.quest.yaml": PASS_QUEST_1},
    )
    pack_file = pack / "pack.yaml"
    updated = pack_file.read_text(encoding="utf-8").replace("Core Wellness Pack", "Core \u2066Wellness Pack")
    pack_file.write_text(updated, encoding="utf-8")
    findings = lint_path(pack, docs_dir=DOCS_DIR)
    assert "SEC-CONTENT-004" in _rules(findings)
